+++ sample-tests/hooks.js
@@ -1,0 +1,19 @@
+(function () {
+
+  'use strict';
+
+  module.exports = function () {
+
+    var library = this;
+
+    library.Before(function (next) {
+      next();
+    });
+
+    library.After(function (next) {
+      next();
+    });
+
+  };
+
+})();