+++ History.md
@@ -1,2 +1,6 @@
+#0.0.2
+Fixed sample test copying
+Hid Before/After steps from showing when they don't error
+
 #0.0.1
 Initial release. Simple cucumberjs + velocity integration